+++ src/main/java/utils/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package utils.booleanGenerator;
+package utils;
 
 public interface BooleanGenerator {
     boolean generate();
